requirement.txt.txt
aniso8601==7.0.0
asgiref==3.5.0
Django==3.0
django-cors-headers==3.11.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-tenant-schemas==1.11.0
django-tenants==3.4.2
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphene-graphiql-explorer==0.0.1
graphql-core==2.3.2
graphql-relay==2.0.1
ordered-set==4.1.0
promise==2.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3